Struct crossbeam_queue::ArrayQueue

pub struct ArrayQueue<T> { /* fields omitted */ }

A bounded multi-producer multi-consumer queue.

This queue allocates a fixed-capacity buffer on construction, which is used to store pushed elements. The queue cannot hold more elements than the buffer allows. Attempting to push an element into a full queue will fail. Having a buffer allocated upfront makes this queue a bit faster than SegQueue.

Examples

use crossbeam_queue::{ArrayQueue, PushError};

let q = ArrayQueue::new(2);

assert_eq!(q.push('a'), Ok(()));
assert_eq!(q.push('b'), Ok(()));
assert_eq!(q.push('c'), Err(PushError('c')));
assert_eq!(q.pop(), Ok('a'));

Methods

impl<T> ArrayQueue<T>

pub fn new(cap: usize) -> ArrayQueue<T>

Creates a new bounded queue with the given capacity.

Panics

Panics if the capacity is zero.

Examples

use crossbeam_queue::ArrayQueue;

let q = ArrayQueue::<i32>::new(100);

pub fn push(&self, value: T) -> Result<(), PushError<T>>

Attempts to push an element into the queue.

If the queue is full, the element is returned back as an error.

Examples

use crossbeam_queue::{ArrayQueue, PushError};

let q = ArrayQueue::new(1);

assert_eq!(q.push(10), Ok(()));
assert_eq!(q.push(20), Err(PushError(20)));

pub fn pop(&self) -> Result<T, PopError>

Attempts to pop an element from the queue.

If the queue is empty, an error is returned.

Examples

use crossbeam_queue::{ArrayQueue, PopError};

let q = ArrayQueue::new(1);
assert_eq!(q.push(10), Ok(()));

assert_eq!(q.pop(), Ok(10));
assert_eq!(q.pop(), Err(PopError));

pub fn capacity(&self) -> usize

Returns the capacity of the queue.

Examples

use crossbeam_queue::{ArrayQueue, PopError};

let q = ArrayQueue::<i32>::new(100);

assert_eq!(q.capacity(), 100);

pub fn is_empty(&self) -> bool

Returns true if the queue is empty.

Examples

use crossbeam_queue::{ArrayQueue, PopError};

let q = ArrayQueue::new(100);

assert!(q.is_empty());
q.push(1).unwrap();
assert!(!q.is_empty());

pub fn is_full(&self) -> bool

Returns true if the queue is full.

Examples

use crossbeam_queue::{ArrayQueue, PopError};

let q = ArrayQueue::new(1);

assert!(!q.is_full());
q.push(1).unwrap();
assert!(q.is_full());

pub fn len(&self) -> usize

Returns the number of elements in the queue.

Examples

use crossbeam_queue::{ArrayQueue, PopError};

let q = ArrayQueue::new(100);
assert_eq!(q.len(), 0);

q.push(10).unwrap();
assert_eq!(q.len(), 1);

q.push(20).unwrap();
assert_eq!(q.len(), 2);